Split ReferenceDataController into multiple controllers
