+++ AdminWebsite/AdminWebsite/Controllers/ReferenceData/ReferenceDataControllerBase.cs
@@ -1,0 +1,8 @@
+using Microsoft.AspNetCore.Mvc;
+
+namespace AdminWebsite.Controllers.ReferenceData;
+
+[Produces("application/json")]
+[Route("api/reference")]
+[ApiController]
+public abstract class ReferenceDataControllerBase : ControllerBase;
